% Generated by roxygen2 (4.0.1): do not edit by hand
\name{almostEqual}
\alias{almostEqual}
\title{Check whether all elements of a numeric vector are equal within machine precision}
\usage{
almostEqual(x, tol = .Machine$double.eps^0.5)
}
\arguments{
  \item{x}{a numeric vector.}

  \item{tol}{tolerance.}
}
\value{
TRUE if all elements of the vector are identical (within
machine precision). FALSE in all other cases, including if
the vector contains any NAs.
}
\description{
Check whether all elements of a numeric vector are equal
within machine precision
}
\note{
This function is based on Hadley and John's answer to
http://stackoverflow.com/q/4752275
}

