% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{epilink}
\alias{epilink}
\alias{epilink-package}
\title{epilink
Infer epigenetic links between genes and their cis-regulatory regions from epigenetic data.}
\description{
epilink Infer epigenetic links between genes and their
cis-regulatory regions from epigenetic data.
}

